+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 Dependency-free, type-safe Dart unit conversion library. Decouple specific unit
 types you need to work with from the rest of your code.
 
+## [Unreleased]
+### Fixed
+- Prevent precision cap breach when performing additions
+
 ## [2.1.0] - 2022-12-11
 ### Added
 - Extensions on the num types to create any measurement
@@ -168,6 +172,7 @@ types you need to work with from the rest of your code.
 ### Added
 - Initial package setup
 
+[Unreleased]: https://bitbucket.org/mongoose13/fling-units/commits/
 [2.1.0]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.1.0/
 [2.0.2]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.0.2/
 [2.0.1]: https://bitbucket.org/mongoose13/fling-units/commits/tag/2.0.1/